Move models module to api module
Resolve clippies
