+++ fpx/src/api/ws.rs
@@ -1,5 +1,5 @@
+use super::models::{ClientMessage, ServerError, ServerMessage, FPX_WEBSOCKET_ID_HEADER};
 use crate::events::ServerEvents;
-use crate::models::{ClientMessage, ServerError, ServerMessage, FPX_WEBSOCKET_ID_HEADER};
 use axum::extract::ws::{Message, WebSocket};
 use axum::extract::{State, WebSocketUpgrade};
 use axum::response::Response;
