+++ fpx/src/commands/debug/ws.rs
@@ -1,4 +1,4 @@
-use crate::models::FPX_WEBSOCKET_ID_HEADER;
+use crate::api::models::FPX_WEBSOCKET_ID_HEADER;
 use anyhow::{Context, Result};
 use futures_util::{SinkExt, StreamExt};
 use std::sync::Arc;
