+++ fpx/src/lib.rs
@@ -2,4 +2,3 @@ pub mod api;
 pub mod data;
 pub mod events;
 pub mod inspector;
-pub mod models;
