+++ fpx/src/main.rs
@@ -16,7 +16,6 @@ mod commands;
 pub mod data;
 mod events;
 mod inspector;
-pub mod models;
 
 #[tokio::main]
 async fn main() -> Result<()> {
